refactor : ProductItemCell 상속 제거 및 Protocol 구현
